refactor: dto 재구성
