--- src/main/java/subway/application/SectionDeleteRequest.java
@@ -1,22 +1,0 @@
-package subway.application;
-
-public class SectionDeleteRequest {
-    private String upStation;
-    private String downStation;
-
-    public SectionDeleteRequest() {
-    }
-
-    public SectionDeleteRequest(final String upStation, final String downStation) {
-        this.upStation = upStation;
-        this.downStation = downStation;
-    }
-
-    public String getUpStation() {
-        return upStation;
-    }
-
-    public String getDownStation() {
-        return downStation;
-    }
-}
