+++ src/main/java/subway/ui/dto/request/SectionDeleteRequest.java
@@ -1,0 +1,16 @@
+package subway.ui.dto.request;
+
+public class SectionDeleteRequest {
+    private String stationName;
+
+    public SectionDeleteRequest() {
+    }
+
+    public SectionDeleteRequest(final String stationName) {
+        this.stationName = stationName;
+    }
+
+    public String getStationName() {
+        return stationName;
+    }
+}
